cut RegularMessageItem into files
